CartView js upgrade
